👌 IMPROVE: Use separate files for dialogs and alerts
